+++ src/FTPClient_Generic.h
@@ -9,11 +9,12 @@
     
   Built by Khoi Hoang https://github.com/khoih-prog/FTPClient_Generic
   
-  Version: 1.0.0
+  Version: 1.1.0
     
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K Hoang      11/05/2022 Initial porting and coding to support many more boards, using WiFi or Ethernet
+  1.1.0   K Hoang      13/05/2022 Add support to Teensy 4.1 using QNEthernet or NativeEthernet
  *****************************************************************************************************************************/
 
 #pragma once
@@ -23,13 +24,13 @@
 
 /////////////////////////////////////////////////////////
 
-#define FTPCLIENT_GENERIC_VERSION            "FTPCLIENT_GENERIC v1.0.0"
+#define FTPCLIENT_GENERIC_VERSION            "FTPCLIENT_GENERIC v1.1.0"
 
 #define FTPCLIENT_GENERIC_VERSION_MAJOR      1
-#define FTPCLIENT_GENERIC_VERSION_MINOR      0
+#define FTPCLIENT_GENERIC_VERSION_MINOR      1
 #define FTPCLIENT_GENERIC_VERSION_PATCH      0
 
-#define FTPCLIENT_GENERIC_VERSION_INT        1000000
+#define FTPCLIENT_GENERIC_VERSION_INT        1001000
 
 /////////////////////////////////////////////////////////